+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 
+* When creating tachyons, include `<tachyon/api/Tachyon.hxx>` to get the templated create code (it's been moved)
 * Global font scaling set to 1.25; which makes it readable but with artifacts on my monitor.  Will need to make this a setting and do it properly (with a font load)
 * Toolbar renamed to button bar (might be further renamed, but wanted it separate to "tools")
 * Reorganization to the tachyon file layout (sorry, getting cluttered again)